% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFTP.R
\name{CFTP}
\alias{CFTP}
\title{Coupling From The Past}
\usage{
CFTP(k, roll.fun, update.fun, monotonic = FALSE, min = NA, max = NA,
  verbose = FALSE, double_time = FALSE, ...)
}
\arguments{
\item{k}{number of states of the Markov chain}

\item{roll.fun}{user defined function that rolls the original die. Must have as first input the size
of the returned sample. Additional inputs can be passed.}

\item{update.fun}{update function of the Markov chain. See the method \code{update.fun} of
class \code{Ladder} to check how it should be defined.}

\item{monotonic}{if \code{TRUE}, monotonic CFTP is implemented}

\item{min}{minimum state used in the monotonic CFTP. Ignored if \code{monotonic = FALSE}.}

\item{max}{maximum state used in the monotonic CFTP. Ignored if \code{monotonic = FALSE}.}

\item{verbose}{if \code{TRUE}, the time steps required by the algorithm is returned as well.}

\item{double_time}{if \code{TRUE} at each iteration of the algorithm, the time is doubled instead
of increased by one. Can lead to slower implementation if rolling the die is computationally costly.}
}
\value{
If \code{verbose=FALSE}, returns a state sampled from the stationary distribution.
If \code{verbose=TRUE}, returns a list where the first element is a state from the stationary
distribution and the second element are the time steps required by the algorithm to finish.
}
\description{
Sample from the stationary distribution of a discrete Markov chain using
the Coupling From the Past algorithm.
}
\details{
The function is designed to sample from a fine and connected ladder using
rolls of a given die and uniform random variables. It requires an update function to be
defined as well as a function to roll the original die. It is mostly used internally by
the \code{Ladder} class.
}
