% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladder.R
\docType{class}
\name{DiscreteDist}
\alias{DiscreteDist}
\title{Categorical distributions}
\format{\code{\link{R6Class}} object.}
\usage{
DiscreteDist
}
\value{
Object of \code{\link{R6Class}} with no methods.
}
\description{
Class to describe cateogrical distributions such as multi-variate ladders.
}
\section{Fields}{

\describe{
\item{\code{k}}{number of categories}

\item{\code{R}}{vector of length \code{k} with the associated probabilities (may not be normalized)}
}}

\keyword{datasets}
